split test utils into separate file
